식당 조회 API Redis 캐시 적용 및 캐싱 전략 개선

+++ src/main/java/com/goorm/tablepick/domain/restaurant/entity/Restaurant.java
@@ -30,7 +30,7 @@ public class Restaurant {
     
     private Double ycoordinate;
     
-    private Long maxCapacity;
+    private int maxCapacity;
 
 
     @OneToMany(mappedBy = "restaurant", cascade = CascadeType.ALL)
